fix(int -> double)

+++ src/binpacking.cpp
@@ -17,10 +17,10 @@ void BinPacking::solve(){
         // std::cout << "vendedor " << j << " deposito " << deposito << std::endl;
 
         if (deposito == -1) {
-            int demanda_maxima_vj = 0;
+            double demanda_maxima_vj = 0;
 
             for (int i = 0; i < this->_instance.m(); i++){
-                int c_dem = this->_instance.demanda(i, j);
+                double c_dem = this->_instance.demanda(i, j);
                 if (c_dem > demanda_maxima_vj){
                     demanda_maxima_vj = c_dem;
                 }
@@ -46,19 +46,18 @@ int BinPacking::get_mejor_deposito(int j) {
     int mejor_costo = -1;
 
     // Busco el deposito i cuya capacidad al restarle la demanda del vendedor j sea la menor
-    
-    std::vector<std::tuple<int, int>> capacidades_con_j(this->_instance.m());
+
+    std::vector<std::tuple<int, double>> capacidades_con_j(this->_instance.m());
 
     for (int deposito = 0; deposito < this->_instance.m(); deposito++) {
-        int cap = this->get_capacidad_deposito(deposito) - this->_instance.demanda(deposito, j);
+        double cap = this->get_capacidad_deposito(deposito) - this->_instance.demanda(deposito, j);
 
         capacidades_con_j[deposito] = std::make_tuple(cap, deposito);
     }
 
     // Ordenar capacidades de mayor a menor por el primer elemento de la tupla
-    std::sort(capacidades_con_j.begin(), capacidades_con_j.end(), [](const std::tuple<int, int> & a, const std::tuple<int, int> & b) -> bool {
-        return std::get<0>(a) > std::get<0>(b);
-    });
+    std::sort(capacidades_con_j.begin(), capacidades_con_j.end(), [](const std::tuple<int, double> &a, const std::tuple<int, double> &b) -> bool
+              { return std::get<0>(a) > std::get<0>(b); });
 
     // Si la capacidad del primer elemento de la lista ordenada es menor a 0, no hay deposito que cumpla con la demanda del vendedor j
     if (std::get<0>(capacidades_con_j[0]) < 0) {
